requirements.txt
adafruit-ampy==1.1.0
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
bitstring==4.0.2
blinker==1.6.2
CacheControl==0.13.1
cachetools==5.3.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==41.0.3
discord.py==2.3.2
distlib==0.3.6
ecdsa==0.18.0
esptool==4.6.2
filelock==3.10.7
firebase-admin==6.2.0
Flask==2.3.2
Flask-CKEditor==0.4.6
Flask-SQLAlchemy==3.0.3
frozenlist==1.3.3
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.13.1
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
greenlet==2.0.2
grpcio==1.59.2
grpcio-status==1.59.2
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
msgpack==1.0.7
multidict==6.0.4
pi==0.1.2
pipenv==2023.3.20
platformdirs==3.2.0
proto-plus==1.22.3
protobuf==4.25.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
pyserial==3.5
python-dotenv==1.0.0
python-firebase==1.2
PyYAML==6.0.1
reedsolo==1.7.0
requests==2.31.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.14
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==2.1.0
virtualenv==20.21.0
virtualenv-clone==0.5.7
Werkzeug==2.3.4
yarl==1.8.2
